% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGompertz.R, R/fit_gompertz.R
\name{fit_gompertz}
\alias{fit_gompertz}
\alias{fit_gompertz}
\title{fit_gompertz is the primary function for fitting growth data}
\usage{
fit_gompertz(dataframe)

fit_gompertz(dataframe)
}
\arguments{
\item{dataframe}{The dataframe with columns: Year,}

\item{dataframe}{The dataframe with columns: Year,}
}
\value{
list

list
}
\description{
fit_gompertz is the primary function for fitting growth data

fit_gompertz is the primary function for fitting growth data
}
